DDD-92 Organize calculators, change CDI scopes
