--- src/main/java/lt/vu/application/order/service/PriceCalculator.java
@@ -1,10 +1,0 @@
-package lt.vu.application.order.service;
-
-import lt.vu.persistence.orm.entities.Order;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-
-public interface PriceCalculator {
-    int calculate(Order order);
-}
