+++ src/main/java/lt/vu/application/order/service/calculator/ChainablePriceCalculator.java
@@ -1,4 +1,4 @@
-package lt.vu.application.order.service;
+package lt.vu.application.order.service.calculator;
 
 import lt.vu.persistence.orm.entities.Order;
