+++ src/main/java/lt/vu/application/order/service/calculator/DefaultPriceCalculator.java
@@ -1,12 +1,13 @@
-package lt.vu.application.order.service;
+package lt.vu.application.order.service.calculator;
 
 import lt.vu.persistence.orm.entities.Order;
 
-import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
-@ApplicationScoped
+@RequestScoped
 public class DefaultPriceCalculator implements PriceCalculator {
+
     @Inject
     private Iterable<ChainablePriceCalculator> calculators;
