+++ src/main/java/lt/vu/application/order/service/calculator/DiscountedPriceCalculator.java
@@ -1,15 +1,15 @@
-package lt.vu.application.order.service;
+package lt.vu.application.order.service.calculator;
 
 import lt.vu.persistence.orm.entities.Order;
 
-import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.context.RequestScoped;
 import javax.enterprise.inject.Alternative;
 import javax.enterprise.inject.Specializes;
 
 @Alternative
 @Specializes
-@ApplicationScoped
-public class DiscountedPriceCalculator extends DefaultPriceCalculator{
+@RequestScoped
+public class DiscountedPriceCalculator extends DefaultPriceCalculator {
 
     private final int DISCOUNT_AMOUNT = 20;
