+++ src/main/java/lt/vu/application/order/service/calculator/FragilePriceCalculator.java
@@ -1,14 +1,14 @@
-package lt.vu.application.order.service;
+package lt.vu.application.order.service.calculator;
 
 import lombok.SneakyThrows;
 import lt.vu.persistence.orm.entities.AttributeType;
 import lt.vu.persistence.orm.entities.Order;
 import lt.vu.persistence.orm.repository.AttributeRepository;
 
-import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
-@ApplicationScoped
+@RequestScoped
 public class FragilePriceCalculator implements ChainablePriceCalculator {
 
     @Inject
