+++ src/main/java/lt/vu/application/order/service/calculator/PriceCalculator.java
@@ -1,0 +1,8 @@
+package lt.vu.application.order.service.calculator;
+
+import lt.vu.persistence.orm.entities.Order;
+
+public interface PriceCalculator {
+
+    int calculate(Order order);
+}
