+++ src/main/java/lt/vu/application/order/service/producer/ChainableCalculatorProducer.java
@@ -1,6 +1,9 @@
-package lt.vu.application.order.service;
+package lt.vu.application.order.service.producer;
 
-import javax.enterprise.context.ApplicationScoped;
+import lt.vu.application.order.service.calculator.ChainablePriceCalculator;
+import lt.vu.application.order.service.calculator.FragilePriceCalculator;
+
+import javax.enterprise.context.RequestScoped;
 import javax.enterprise.inject.Produces;
 import javax.inject.Inject;
 import java.util.ArrayList;
@@ -9,7 +12,7 @@ import java.util.Arrays;
 /**
  * Produces a list of ChainablePriceCalculator implementations for injection purposes.
  */
-@ApplicationScoped
+@RequestScoped
 public class ChainableCalculatorProducer {
 
     @Inject
