+++ src/main/webapp/WEB-INF/beans.xml
@@ -10,6 +10,6 @@
         <class>lt.vu.infrastructure.interceptors.MethodLogger</class>
     </interceptors>
     <alternatives>
-        <class>lt.vu.application.order.service.DiscountedPriceCalculator</class>
+        <class>lt.vu.application.order.service.calculator.DiscountedPriceCalculator</class>
     </alternatives>
 </beans>
